move video writer to own function
